requirements.txt
click==8.1.3
colorama==0.4.6
Flask==2.2.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
Werkzeug==2.2.3
azure-keyvault-secrets
azure-identity
flask-sqlalchemy
psycopg2